requirements.txt
python-dotenv==1.0.0
rich==13.5.2
requests==2.31.0